requirements.txt
alembic==1.2.0
asn1crypto==0.24.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.1
blinker==1.4
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.7
confusable-homoglyphs==3.2.0
cryptography==2.1.4
Django==1.11.26
django-bootstrap4==1.0.1
django-registration==2.4.1
django-tinymce==2.8.0
dominate==2.4.0
enum34==1.1.6
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
futures==3.3.0
google-assistant-grpc==0.2.1
google-assistant-sdk==0.5.1
google-auth==1.6.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gyp==0.1
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
oauthlib==3.1.0
pathlib2==2.3.4
Pillow==6.2.1
protobuf==3.9.1
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-decouple==3.3
python-editor==1.0.4
pytz==2019.3
pyxdg==0.25
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scandir==1.10.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sounddevice==0.3.13
soupsieve==1.9.5
SQLAlchemy==1.3.8
tenacity==4.12.0
urllib3==1.25.3
virtualenv==16.7.7
visitor==0.1.3
Werkzeug==0.16.0
